requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
decorator==4.3.0
Django==1.11
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
imageio==2.3.0
jwt==0.5.4
moviepy==0.2.3.5
mutagen==1.41.0
numpy==1.14.5
Pillow==5.2.0
pycparser==2.18
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
six==1.11.0
tqdm==4.23.4
urllib3==1.23
